refactor(widgets): remove props elements
