--- src/shared/ui/carousel/types.ts
@@ -1,9 +1,0 @@
-import { EmblaCarouselType } from "embla-carousel";
-
-export type CarouselControlsProps = {
-  emblaApi?: EmblaCarouselType;
-};
-
-export type CarouselProps = {
-  children: React.ReactElement[];
-};
