+++ src/widgets/coaches/ui/coach-сard.tsx
